+++ lib/features/cart/widgets/cart_subtotal.dart
@@ -11,7 +11,7 @@ class CartSubtotal extends StatelessWidget {
     final user = context.watch<UserProvider>().user;
     double sum = 0;
     user.cart
-        .map((e) => sum += e['quantity'] * e['product']['discountPrice'] as double)
+        .map((e) => sum += e['quantity'] * e['product']['discountPrice'])
         .toList();
 
     return Container(
